compare back&front

+++ WebBack/WebBack/Data/Entities/EngineVolumeEntity.cs
@@ -5,7 +5,7 @@ namespace WebBack.Data.Entities
     [Table("tbl_engine_volumes")]
     public class EngineVolumeEntity : BaseEntity
     {
-        public decimal Volume { get; set; }
+        public float Volume { get; set; }
 
 
     }
